Changing component name to match return
